Agrego opcion para hacer gráficos solo de puntos a linePlot Agrego controles basados en media y SD Agrego la máscara de ceros a los mapas de diagnóstico Agrego soporte a escalas con transparencia

+++ SeriesTemporales/leerSeriesTemporales.r
@@ -128,7 +128,7 @@ leerSeriesXLSX <- function(pathArchivoDatos, hojaEstaciones='InfoPluvios', heade
 
 grabarSeriesArchivoUnico <- function(pathArchivoDatos, estaciones, fechas, datos, sep='\t', dec='.', na='-9999',
                                      colsEstacionesAEscribir=1:ncol(estaciones), formatoFechas='%Y-%m-%d') {
-  unlink(pathArchivoDatos)
+  if (file.exists(pathArchivoDatos)) unlink(pathArchivoDatos)
   
   # i <- colsEstacionesAEscribir[1]
   for (i in colsEstacionesAEscribir) {
